% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_clCode}
\alias{check_clCode}
\title{Check HS classification parameter}
\usage{
check_clCode(commodity_classification)
}
\arguments{
\item{commodity_classification}{A character string specifying the HS classification code. Must be "HS".}
}
\value{
A character string specifying the HS classification code.
}
\description{
Check HS classification parameter
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
untrader:::check_clCode("HS") # returns "HS"
untrader:::check_clCode("ISIC") # throws an error because "ISIC" is not a valid classification code
\dontshow{\}) # examplesIf}
}
