% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_date}
\alias{check_date}
\title{Check date parameter}
\usage{
check_date(start_date, end_date, frequency)
}
\arguments{
\item{start_date}{The start date of the query, either in the format `yyyy` or `yyyy-mm`.}

\item{end_date}{The end date of the query, either in the format `yyyy` or `yyyy-mm`. Can be a maximum of 12 years after the start date for the annuel frequency or one year for monthly.}

\item{frequency}{The frequency of reported trade data, either `A` for annual or `M` for monthly.}
}
\value{
A character vector of valid reporter IDs.
}
\description{
This function checks that the given period code is valid. If the range or format is not valid, the function throws an error message indicating which codes are invalid. It also converts the input to the proper format if necessary.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
check_date(2010,2011,'A')
\dontshow{\}) # examplesIf}
}
