% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_flowCode}
\alias{check_flowCode}
\title{Check validity of flow direction parameter.}
\usage{
check_flowCode(flow_direction)
}
\arguments{
\item{flow_direction}{The direction of trade flows, e.g. whether you would like to get data on reported imports or exports. Possible values are "import" for imports, "export" for exports, "re-import" for re-imports, "re-export" for re-exports, or "all" for imports, exports, re-imports, and re-exports. Multiple values can be supplied as a character vector. The default value is "all".}
}
\value{
A character vector specifying the trade flow codes.
}
\description{
Trade flow code: export, import, re-export, re-import.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
check_flowCode("import") # returns "M"
check_flowCode(c("export", "re-export")) # returns "X,RX"
check_flowCode("trade") # throws an error because "trade" is not a valid flow code
check_flowCode(NULL) # throws an error because at least one flow code must be provided
\dontshow{\}) # examplesIf}
}
