% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_partnerCode}
\alias{check_partnerCode}
\title{Check partner parameter}
\usage{
check_partnerCode(partner)
}
\arguments{
\item{partner}{A character vector or string of comma-separated codes that
represent the trade partners in the trade data. The default value is NULL.}
}
\value{
A character vector of valid partner IDs.
}
\description{
This function checks that the given partner code is valid. If the code is not
valid, the function throws an error message indicating which codes are invalid.
It also converts the input to a proper format if necessary.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
check_partnerCode("CAN")
check_partnerCode(c("CAN", "MEX"))
check_partnerCode("all")
\dontshow{\}) # examplesIf}
}
