% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtradr-package.R
\docType{package}
\name{comtradr-package}
\alias{comtradr}
\alias{comtradr-package}
\title{comtradr: Interface with the United Nations Comtrade API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Interface with and extract data from the United Nations 'Comtrade' API \url{https://comtradeplus.un.org/}. 'Comtrade' provides country level shipping data for a variety of commodities, these functions allow for easy API query and data returned as a tidy data frame.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/comtradr/}
  \item \url{https://github.com/ropensci/comtradr}
  \item Report bugs at \url{https://github.com/ropensci/comtradr/issues}
}

}
\author{
\strong{Maintainer}: Paul Bochtler \email{paulbochtler.gh@gmail.com} (\href{https://orcid.org/0000-0002-9146-6185}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Harriet Goers \email{hgoers@umd.edu}
  \item Chris Muir \email{chrismuirRVA@gmail.com}
}

Other contributors:
\itemize{
  \item Alicia Schep (\href{https://orcid.org/0000-0002-3915-0618}{ORCID}) (Alicia reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/141) [reviewer]
  \item Rafael Hellwig (\href{https://orcid.org/0000-0002-3092-3493}{ORCID}) (Rafael reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/141) [reviewer]
  \item Ernest Guevarra (\href{https://orcid.org/0000-0002-4887-4415}{ORCID}) (Ernest reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/613) [reviewer]
  \item Nicholas Potter (\href{https://orcid.org/0000-0002-3410-3732}{ORCID}) (Nicholas reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/613) [reviewer]
  \item Juergen Amann [contributor]
}

}
\keyword{internal}
