% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{Country codes}
\format{
`country_codes`
A dataframe with 312 rows and seven columns:
\describe{
 \item{id}{Unique country code.}
 \item{country}{Name of the country (in English).}
 \item{iso_3}{The country's ISO 3 code.}
 \item{entry_year}{The country's entry into the international system or 1900 (whichever is largest).}
 \item{exit_year}{The country's exit from the international system, if applicable.}
 \item{reporter}{Indicates whether the country is a reporter in the UN Comtrade database.}
 \item{partner}{Indicates whether the country can be reported on by others in the UN Comtrade database. Not all partners are reporters. For example, the World cannot report its trade values.}
}
}
\source{
<https://comtradeapi.un.org/files/v1/app/reference/Reporters.json> and <https://comtradeapi.un.org/files/v1/app/reference/partnerAreas.json>
}
\usage{
country_codes
}
\description{
A full dataset of all reporter and partner codes available in the UN Comtrade database.
}
\keyword{datasets}
