% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_build_request.R
\name{ct_build_request}
\alias{ct_build_request}
\title{Build a valid request object from the checked parameters}
\usage{
ct_build_request(params, primary_token = NULL, verbose = FALSE)
}
\arguments{
\item{params}{a named vector of parameters for the comtrade request}

\item{primary_token}{Your primary token. Default is to check in environment for stored token, if not passed through the \code{set_primary_comtrade_key} function}

\item{verbose}{whether the function sends status updates to the console}
}
\value{
a httr2 request object
}
\description{
This function takes the necessary parameters and creates a httr2 request to be performed this request can then be used in a second function, to actually return the data
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Build request from checked parameters
ct_build_request(comtradr:::ct_check_params(type = 'goods',
                                            frequency = 'A',
                                            commodity_classification = 'HS',
                                            commodity_code = 'TOTAL',
                                            flow_direction = 'all',
                                            reporter = 'all',
                                            partner = 'World',
                                            start_date = 2020,
                                            end_date = 2022,
                                            process = TRUE,
                                            tidy_cols = TRUE,
                                            verbose = FALSE,
                                            primary_token = 'xxxx',
                                            mode_of_transport = '0',
                                            partner_2 = 'World',
                                            customs_code ='C00',
                                            update = FALSE ))
\dontshow{\}) # examplesIf}
}
