% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_commodity_lookup}
\alias{ct_commodity_lookup}
\title{ct_commodity_lookup}
\usage{
ct_commodity_lookup(...)
}
\arguments{
\item{...}{Used to catch all possible arguments that users have supplied to this function.}
}
\description{
This function is deprecated. Please use  \code{xyz()} instead.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
