% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ct_get_commodity_table}
\alias{ct_get_commodity_table}
\title{Get reference table from package data}
\usage{
ct_get_commodity_table(commodity_classification)
}
\arguments{
\item{commodity_classification}{The trade (IMTS) classification scheme. A character value. The only possible value is 'HS'. This is the default.}
}
\description{
Get reference table from package data
}
