% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limit.R
\name{ct_get_remaining_hourly_queries}
\alias{ct_get_remaining_hourly_queries}
\title{Comtradr rate limit check}
\usage{
ct_get_remaining_hourly_queries()
}
\value{
numeric value, number of current queries left in the hour.
}
\description{
Get the remaining number of queries left in the current hour.
}
\examples{
ct_get_remaining_hourly_queries()
}
