% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limit.R
\name{ct_get_reset_time}
\alias{ct_get_reset_time}
\title{Comtradr rate limit time check}
\usage{
ct_get_reset_time(set = NULL)
}
\arguments{
\item{set}{logical, if \code{TRUE} and the current reset time is
\code{NULL}, set the reset time to be one hour from the current
\code{Sys.time}.}
}
\value{
date and time in which the hourly query limit will reset. Return is
  a "POSIXct" object (see \code{\link{DateTimeClasses}}).
}
\description{
Get the time in which the hourly limit will reset.
}
\examples{
ct_get_reset_time()

# Get minutes remaining until limit reset, as numeric value.
as.double(ct_get_reset_time() - Sys.time())
}
