% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ct_migrate_cache}
\alias{ct_migrate_cache}
\title{Migrate cache to new location}
\usage{
ct_migrate_cache()
}
\value{
Nothing
}
\description{
Comtradr versions previous to version 1.0.1 have used a cache
location that was not CRAN compliant. You can migrate any remaining files to
the new cache location using this function. It will delete the old cache.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ct_migrate_cache()
\dontshow{\}) # examplesIf}
}
