% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct_pretty_cols}
\alias{ct_pretty_cols}
\title{ct_pretty_cols}
\format{
\code{country_codes}
A dataframe with 47 rows and twi columns:
\describe{
\item{to}{tidy columns}
\item{from}{original column names}
}
}
\usage{
ct_pretty_cols
}
\description{
A data.frame with a matched list of tidy and untidy column names for the results.
}
\keyword{datasets}
