% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_search}
\alias{ct_search}
\title{ct_search}
\usage{
ct_search(...)
}
\arguments{
\item{...}{Used to catch all possible arguments that users have supplied to this function.}
}
\description{
This function is deprecated Please use  `ct_get_data()` instead.
`r lifecycle::badge("superseded")`
}
