% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_update_databases}
\alias{ct_update_databases}
\title{ct_update_databases}
\usage{
ct_update_databases()
}
\description{
This function is deprecated Please use  `xyz()` instead.
`r lifecycle::badge("superseded")`
}
