% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_use_pretty_cols}
\alias{ct_use_pretty_cols}
\title{ct_use_pretty_cols}
\usage{
ct_use_pretty_cols(...)
}
\arguments{
\item{...}{Used to catch all possible arguments that users have supplied to this function.}
}
\value{
depreciation error
}
\description{
This function is deprecated. Please use the \code{process} argument in the main function instead.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
