% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_search.R
\name{execute_api_request}
\alias{execute_api_request}
\title{Send API request to Comtrade.}
\usage{
execute_api_request(url, col_name)
}
\arguments{
\item{url}{str, url to send to the Comtrade API.}

\item{col_name}{str, indicating whether to edit the col headers of the
return data frame.}
}
\value{
data frame of API return data.
}
\description{
Send API request to Comtrade.
}
