% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_primary_comtrade_key}
\alias{get_primary_comtrade_key}
\title{get_primary_comtrade_key}
\usage{
get_primary_comtrade_key()
}
\value{
Gets your primary comtrade key from the environment var COMTRADE_PRIMARY
}
\description{
If you would like your Comtrade API key to persist in between sessions, use `usethis::edit_r_environ()` to add the env variable COMTRADE_PRIMARY to your environment file.
}
