% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_response.R
\name{process_comtrade_response}
\alias{process_comtrade_response}
\title{Processes the response object}
\usage{
process_comtrade_response(resp, verbose = F)
}
\arguments{
\item{resp}{a valid httr2 response object created from the function `perform_comtrade_request`}

\item{verbose}{whether the function sends status updates to the console}
}
\value{
a data.frame object with the results
}
\description{
The function processes the httr2response object, parses the json and adds the respective iso codes for the reporter and partner countries, as well as the commodity code description.
}
