% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_cmdCode}
\alias{check_cmdCode}
\title{Check commodity_code parameter}
\usage{
check_cmdCode(commodity_code)
}
\arguments{
\item{commodity_code}{A character string or vector specifying the HS codes.}
}
\value{
A character vector specifying the HS codes.
}
\description{
Check commodity_code parameter
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
check_cmdCode("01") # returns "01"
check_cmdCode(c("01", "02")) # returns "01,02"
check_cmdCode("ABC") # throws an error because "ABC" is not a valid HS code
check_cmdCode(NULL) # throws an error because at least one HS code must be provided
\dontshow{\}) # examplesIf}
}
