% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_reporterCode}
\alias{check_reporterCode}
\title{Check reporter parameter}
\usage{
check_reporterCode(reporter)
}
\arguments{
\item{reporter}{A character vector or string of comma-separated codes that
represent the reporters in the trade data. The default value is NULL.}
}
\value{
A character vector of valid reporter IDs.
}
\description{
This function checks that the given reporter code is valid. If the code is not
valid, the function throws an error message indicating which codes are invalid.
It also converts the input to a proper format if necessary.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
check_reporterCode("USA")
check_reporterCode(c("USA", "FRA"))
check_reporterCode("all")
\dontshow{\}) # examplesIf}
}
