% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmd_codes}
\alias{cmd_codes}
\title{Commodity codes}
\format{
`cmd_codes`
A dataframe with 8,267 rows and four columns:
\describe{
 \item{id}{Unique commodity code.}
 \item{text}{Description of the commodity.}
 \item{parent}{The parent commodity code.}
 \item{class_code}{The commodity's classification code. We currently only support HS codes.}
}
}
\source{
<https://comtrade.un.org/Data/cache/classificationHS.json>
}
\usage{
cmd_codes
}
\description{
A full dataset of all commodity codes available in the UN Comtrade database.
}
\keyword{datasets}
