% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_commodity_db_type}
\alias{ct_commodity_db_type}
\title{ct_commodity_db_type}
\usage{
ct_commodity_db_type(...)
}
\arguments{
\item{...}{Used to catch all possible arguments that users have supplied to this function.}
}
\description{
This function is deprecated. There is currently no alternative for this function.
`r lifecycle::badge("superseded")`
}
