% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_commodity_db_type}
\alias{ct_commodity_db_type}
\title{ct_commodity_db_type}
\usage{
ct_commodity_db_type()
}
\description{
This function is deprecated Please use  `xyz()` instead.
`r lifecycle::badge("superseded")`
}
