% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_commodity_lookup.R
\name{ct_commodity_lookup}
\alias{ct_commodity_lookup}
\title{UN Comtrade commodities database query}
\usage{
ct_commodity_lookup(values, return_code = FALSE, return_char = FALSE,
  verbose = TRUE)
}
\arguments{
\item{values}{Commodity names or commodity codes, as a char or numeric
vector.}

\item{return_code}{Logical, if set to FALSE, the function will return a
set of commodity descriptions along with commodity codes (as a single
string for each match found), if set to TRUE it will return only the
commodity codes. Default value is FALSE.}

\item{return_char}{Logical, if set to FALSE, the function will return the
matches as a named list, if set to TRUE it will return them as a character
vector. Default value is FALSE.}

\item{verbose}{Logical, if set to TRUE, a warning message will print to
console if any of the elements of input "values" returned no matches
(message will indicate which elements returned no data). Default is TRUE.}
}
\value{
A list or character vector of commodity descriptions and/or
 commodity codes that are matches with the elements of "values".
}
\description{
The Comtrade API requires that searches for specific commodities be done
using commodity codes. This is a helper function for querying the
Comtrade commodity database. It takes as input a vector of
commodities or commodity codes. Output is a list or vector of commodity
descriptions or codes associated with the input values. For use with the UN
Comtrade API, full API docs can be found at
\url{https://comtrade.un.org/data/doc/api/}
}
\examples{
\dontrun{
# Look up commodity descriptions related to "halibut"
ct_commodity_lookup("halibut",
                    commoditydf,
                    return_code = FALSE,
                    return_char = FALSE,
                    verbose = TRUE)

# Look up commodity codes related to "shrimp".
ct_commodity_lookup("shrimp",
                    commoditydf,
                    return_code = TRUE,
                    return_char = FALSE,
                    verbose = TRUE)
}
}
