% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_country_lookup}
\alias{ct_country_lookup}
\title{ct_country_lookup}
\usage{
ct_country_lookup(...)
}
\description{
This function is deprecated Please use  `xyz()` instead.
`r lifecycle::badge("superseded")`
}
