% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_country_lookup}
\alias{ct_country_lookup}
\title{ct_country_lookup}
\usage{
ct_country_lookup(...)
}
\arguments{
\item{...}{Used to catch all possible arguments that users have supplied to this function.}
}
\description{
This function is deprecated. You can `use country_codes` to return a dataset with all possible country codes, but in general the specification of iso 3 codes makes a look-up unnecessary.
`r lifecycle::badge("superseded")`
}
