% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_get_data.R
\name{ct_get_data}
\alias{ct_get_data}
\title{Get trade data from the UN Comtrade API}
\usage{
ct_get_data(
  type = "goods",
  frequency = "A",
  commodity_classification = "HS",
  commodity_code = "TOTAL",
  flow_direction = "all",
  reporter = "all",
  partner = "World",
  start_date = NULL,
  end_date = NULL,
  process = TRUE,
  tidy_cols = TRUE,
  verbose = FALSE,
  primary_token = get_primary_comtrade_key(),
  mode_of_transport = "0",
  partner_2 = "World",
  customs_code = "C00",
  update = FALSE,
  ...
)
}
\arguments{
\item{type}{The type of returned trade data. Possible values: 'goods' for trade in goods, 'services' for trade in services. Default: 'goods'.}

\item{frequency}{The frequency of returned trade data. Possible values: 'A' for annual data, 'M' for monthly data. Default: 'A'.}

\item{commodity_classification}{The trade classification scheme. Possible values for goods: \code{c('HS','S1','S2','S3','S4','SS')}; for services: \code{c('B4','B5','EB02','EB10','EB10S','EB')}. Default: 'HS'.}

\item{commodity_code}{The commodity code(s) or \code{NULL}. See \code{comtradr::ct_get_ref_table('HS')} for possible values. Default: 'TOTAL' (sum of all commodities).}

\item{flow_direction}{The direction of trade flows or \code{NULL}. Possible values: 'import', 'export', 're-import', 're-export', 'all'. Default: 'all'.}

\item{reporter}{Reporter ISO3 code(s) or \code{NULL}. See \code{comtradr::country_codes} for possible values. Default: 'all'.}

\item{partner}{Partner ISO3 code(s) or \code{NULL}. See \code{comtradr::country_codes} for possible values. Default: 'World' (all partners as an aggregate).}

\item{start_date}{The start date of the query. Format: \code{yyyy} for yearly, \code{yyyy-mm} for monthly.}

\item{end_date}{The end date of the query. Format: \code{yyyy} for yearly, \code{yyyy-mm} for monthly. Max: 12 years after start date for annual data, one year for monthly data.}

\item{process}{If TRUE, returns a data.frame with results. If FALSE, returns the raw httr2 request. Default: TRUE.}

\item{tidy_cols}{If TRUE, returns tidy column names. If FALSE, returns raw column names. Default: TRUE.}

\item{verbose}{If TRUE, sends status updates to the console. If FALSE, runs functions quietly. Default: FALSE.}

\item{primary_token}{Your primary UN Comtrade API token. Default: stored token from \code{comtradr::set_primary_comtrade_key}.}

\item{mode_of_transport}{Mode of Transport, default '0' (TOTAL). See \code{ct_get_ref_table(dataset_id = 'mot')} for possible values.}

\item{partner_2}{Partner ISO3 code(s) or \code{NULL}. Default: 'World'.}

\item{customs_code}{Customs code, default 'C00' (TOTAL). See \code{ct_get_ref_table(dataset_id = 'customs')} for possible values.}

\item{update}{If TRUE, downloads possibly updated reference tables from the UN. Default: FALSE.}

\item{...}{Additional parameters to the API, passed as query parameters without checking.}
}
\value{
A data.frame with trade data or, if \code{process = F}, a httr2 response object.
}
\description{
This function queries the UN Comtrade API to retrieve international trade data.
It allows for detailed specification of the query, including the type of data (goods or services),
frequency (annual or monthly), commodity classification, flow direction, and more.
By providing \code{NULL} for certain parameters, you can query all possible values.
The function is opinionated in that it already verifies certain parameters for you and
is more than a pure wrapper around the API.
}
\details{
The UN Comtrade database provides a repository of official international trade statistics
and relevant analytical tables. It contains annual trade statistics starting from 1988
and monthly trade statistics since 2000 for goods data

Parameters that accept \code{NULL} will query all possible values. For example, setting \code{commodity_code = NULL}
will retrieve data for all commodity codes. This can be useful for broad queries but may result in large datasets.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Query goods data for China's trade with Argentina and Germany in 2019
ct_get_data(type = 'goods',
            commodity_classification = 'HS',
            commodity_code = 'TOTAL',
            reporter = 'CHN',
            partner = c('ARG','DEU'),
            start_date = '2019',
            end_date = '2019',
            flow_direction = 'all',
            partner_2 = 'World',
            verbose = TRUE)

# Query all commodity codes for China's imports from Germany in 2019
ct_get_data(commodity_code = NULL,
            reporter = 'CHN',
            partner = 'DEU',
            start_date = '2019',
            end_date = '2019',
            flow_direction = 'import')

# Query all commodity codes for China's imports from Germany from January to June of 2019
ct_get_data(commodity_code = NULL,
            reporter = 'CHN',
            partner = 'DEU',
            start_date = '2019',
            end_date = '2019',
            flow_direction = 'import')
\dontshow{\}) # examplesIf}
}
