% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ct_get_ref_table}
\alias{ct_get_ref_table}
\title{Get reference table from package data}
\usage{
ct_get_ref_table(dataset_id, update = F, verbose = F)
}
\arguments{
\item{dataset_id}{The dataset ID, which is either partner, reporter or a valid classification scheme.}
}
\description{
Get reference table from package data
}
