% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ct_get_ref_table}
\alias{ct_get_ref_table}
\title{Get reference table from package data}
\usage{
ct_get_ref_table(dataset_id, update = FALSE, verbose = FALSE)
}
\arguments{
\item{dataset_id}{The dataset ID, which is either partner, reporter or a valid classification scheme.}

\item{update}{A logical value. If TRUE, will download the possibly updated reference tables from the UN.}

\item{verbose}{A logical value. If TRUE, sends status updates to the console. If FALSE, runs functions quietly.}
}
\description{
The first time, the function will read from disk, the second time from the environment. In the case of a necessary update the new data will be saved to the environment for the current session.
}
