% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_get_remaining_hourly_queries}
\alias{ct_get_remaining_hourly_queries}
\title{ct_get_remaining_hourly_queries}
\usage{
ct_get_remaining_hourly_queries(...)
}
\arguments{
\item{...}{Used to catch all possible arguments that users
have supplied to this function.}
}
\value{
depreciation error
}
\description{
This function is deprecated. There is no more reset
time, as the upper limit of 250 calls per day is enforced daily.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# no examples because only legacy function
\dontshow{\}) # examplesIf}
}
