% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_get_reset_time}
\alias{ct_get_reset_time}
\title{ct_get_reset_time}
\usage{
ct_get_reset_time()
}
\description{
This function is deprecated Please use  `xyz()` instead.
`r lifecycle::badge("superseded")`
}
