% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_get_reset_time}
\alias{ct_get_reset_time}
\title{ct_get_reset_time}
\usage{
ct_get_reset_time(...)
}
\arguments{
\item{...}{Used to catch all possible arguments that users have supplied to this function.}
}
\description{
This function is deprecated. There is no more reset time, as the upper limit of 250 calls per day is enforced daily.
`r lifecycle::badge("superseded")`
}
