% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_perform_request.R
\name{ct_perform_request}
\alias{ct_perform_request}
\title{Performs the request to the Comtrade API}
\usage{
ct_perform_request(req, requests_per_second = 10/60, verbose = F)
}
\arguments{
\item{req}{a valid comtrade request built by the `ct_build_request()` function}

\item{requests_per_second}{rate at which throttling is done}

\item{verbose}{whether the function sends status updates to the console}
}
\value{
json data from comtrade and possible error codes
}
\description{
Performs the request and returns an error body with the respective error returned by the Comtrade API. Also throttles all requests to 1 request per 6 seconds, or 10 requests per minute.
}
