% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_process_response.R
\name{ct_process_response}
\alias{ct_process_response}
\title{Processes the response object}
\usage{
ct_process_response(resp, verbose = FALSE, tidy_cols)
}
\arguments{
\item{resp}{a valid httr2 response object created from the function \code{ct_perform_request()}}

\item{verbose}{whether the function sends status updates to the console}

\item{tidy_cols}{A logical value. If TRUE, returns a data.frame with tidy column names. If FALSE, returns the raw column names Defaults to TRUE.}
}
\value{
a data.frame object with the results
}
\description{
The function processes the httr2response object, parses the json and adds the respective iso codes for the reporter and partner countries, as well as the commodity code description.
}
