% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_register_token}
\alias{ct_register_token}
\title{ct_get_remaining_secondly_queries}
\usage{
ct_register_token()
}
\description{
This function is deprecated Please use  `xyz()` instead.
`r lifecycle::badge("superseded")`
}
