% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_register_token}
\alias{ct_register_token}
\title{ct_register_token}
\usage{
ct_register_token(...)
}
\arguments{
\item{...}{Used to catch all possible arguments that users
have supplied to this function.}
}
\value{
depreciation error
}
\description{
This function is deprecated.
Please use  \code{set_primary_comtrade_key()} instead.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# no examples because only legacy function
\dontshow{\}) # examplesIf}
}
