% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_update_databases}
\alias{ct_update_databases}
\title{ct_update_databases}
\usage{
ct_update_databases(...)
}
\arguments{
\item{...}{Used to catch all possible arguments that users
have supplied to this function.}
}
\value{
depreciation error
}
\description{
This function is deprecated. Please use  \code{update}
parameter in the main \code{ct_get_data} function instead.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# no examples because only legacy function
\dontshow{\}) # examplesIf}
}
