% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ct_use_pretty_cols}
\alias{ct_use_pretty_cols}
\title{ct_use_pretty_cols}
\usage{
ct_use_pretty_cols(...)
}
\arguments{
\item{...}{Used to catch all possible arguments that users have supplied to this function.}
}
\description{
This function is deprecated. Please use the `process` argument in the main function instead.
`r lifecycle::badge("superseded")`
}
